Strip out enough citation stuff to be actually reading in the ctags file.
This is working really well and it reads the files extremely fast. The
Fuzzy search is something that I may revisit to see if I can do a non fuzzy
search as well since when you get into very large ctags files like 100k entries
then it is really slow to use the fuzzy search.